backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
pydantic==2.8.2
requests==2.32.3
SQLAlchemy==2.0.36
aiofiles==23.2.1